+++ jest.config.js
@@ -1,12 +1,8 @@
 module.exports = {
     clearMocks: true,
-    moduleFileExtensions: [
-        "ts",
-        "js",
-    ],
+    moduleFileExtensions: ['ts', 'js'],
     preset: 'ts-jest',
     testEnvironment: 'node',
-    testMatch: [
-        "**/*.test.ts"
-    ],
+    testMatch: ['**/*.test.ts'],
+    testTimeout: 150000,
 };