+++ src/config/polaris-server-config.ts
@@ -1,22 +1,30 @@
 import { RealitiesHolder } from '@enigmatis/polaris-common';
-import { PolarisGraphQLLogger } from '@enigmatis/polaris-graphql-logger';
-import { ApplicationProperties, LoggerConfiguration } from '@enigmatis/polaris-logs';
-import { Connection } from '@enigmatis/polaris-typeorm';
+import {
+    AbstractPolarisLogger,
+    ApplicationProperties,
+    LoggerConfiguration,
+} from '@enigmatis/polaris-logs';
+import { PolarisConnection } from '@enigmatis/polaris-typeorm';
 import { ApolloServerExpressConfig } from 'apollo-server-express';
 import { DocumentNode } from 'graphql';
 import { IResolvers } from 'graphql-tools';
+import { ExpressContext } from '..';
 import { MiddlewareConfiguration } from '../index';
+import { SnapshotConfiguration } from './snapshot-configuration';
 
 export interface PolarisServerConfig extends ApolloServerExpressConfig {
     typeDefs: DocumentNode | DocumentNode[] | string | string[];
     resolvers: IResolvers | IResolvers[];
     port: number;
     applicationProperties: ApplicationProperties;
-    logger: LoggerConfiguration | PolarisGraphQLLogger;
+    logger: LoggerConfiguration | AbstractPolarisLogger;
     middlewareConfiguration: MiddlewareConfiguration;
     allowSubscription: boolean;
     customMiddlewares?: any[];
-    customContext?: (context: any) => any;
-    connection?: Connection;
+    customContext?: (context: ExpressContext) => any;
+    connection?: PolarisConnection;
     supportedRealities?: RealitiesHolder;
+    shouldAddWarningsToExtensions: boolean;
+    allowMandatoryHeaders: boolean;
+    snapshotConfig: SnapshotConfiguration;
 }