+++ src/index.ts
@@ -1,8 +1,10 @@
-export { PolarisServer } from './server/polaris-server';
-export { getPolarisContext } from './server/context-creator';
+export { ExpressContext } from './server/express-context';
+export { PolarisServer, app } from './server/polaris-server';
 export { PolarisServerOptions } from './config/polaris-server-options';
 export { MiddlewareConfiguration } from './config/middleware-configuration';
+export { SnapshotConfiguration } from './config/snapshot-configuration';
 export { formatError } from './errors/error-formatter';
+export * from 'apollo-server-errors';
 export { PubSub, gql } from 'apollo-server-express';
 export * from '@enigmatis/polaris-common';
 export { Reality } from '@enigmatis/polaris-common';
@@ -12,3 +14,12 @@ export * from '@enigmatis/polaris-middlewares';
 export * from '@enigmatis/polaris-schema';
 export * from '@enigmatis/polaris-typeorm';
 export { Entity } from '@enigmatis/polaris-typeorm';
+export {
+    IResolvers,
+    IResolverObject,
+    IResolverOptions,
+    IResolversParameter,
+    IResolverValidationOptions,
+    IFieldResolver,
+} from 'graphql-tools';
+export * from './snapshot/paginated-resolver';